requirements.txt
crawl4ai
llama-index
llama-index-vector-stores-chroma
llama-index-llms-openrouter
llama-index-embeddings-openai
llama-index-llms-deepseek
lama-index-embeddings-huggingface
sentence-transformers
chromadb
playwright
python-dotenv
streamlit
